Refactoring to separate those routes which need a file system
